+++ tests/integration/profiles/valid/hash-context.br.js
@@ -1,0 +1,1 @@
+var expectedMd5 = choose({ encoding: "hex" }, "d41d8cd98f00b204e9800998ecf8427e");